BackEnd/requirement.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-oauth-toolkit==1.5.0
django-storages==1.11.1
djangorestframework==3.12.4
drf-yasg==1.20.0
dropbox==11.16.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.9.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.0
permission==0.4.1
Pillow==8.3.1
ply==3.11
promise==2.3
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
stone==3.2.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
